Roles

+++ app/Http/Controllers/JobController.php
@@ -2,30 +2,36 @@
 
 namespace App\Http\Controllers;
 use App\Job;
+use App\Review;
 use App\Http\Requests\CreateJobRequest;
+use jeremykenedy\LaravelRoles\Models\Role;
+use jeremykenedy\LaravelRoles\Models\Permission;
 use Illuminate\Http\Request;
 
 class JobController extends Controller
 {
-   public function index(){
-   		$job = Job::orderBy('id','desc')->get();
-   		$job->load('user');
+   public function index()
+   {
+   		$job = Job::orderBy('id','desc')->paginate(10);
+   		$job->load('user','category');
    		return $job;
    }
 
    public function show (Job $job){
-   		$job->load('user');
+   		$job->load('user','category','reviews');
 
    		return $job;		
    }
 
-   public function store(CreateJobRequest $request){
-   		try {
+   public function store(CreateJobRequest $request)
+   {	 
+         try {
    			$job = Job::create([
    				'name' => $request->input('name'),
    				'description' => $request->input('description'),
    				'price' => $request->input('price'),
-   				'user_id' => '5'
+   				'user_id' => $request->user()->id,
+               'category_id' => $request->input('category')
    			]);
 
    		$response['status'] = true;
@@ -37,19 +43,34 @@ class JobController extends Controller
     		$response['status'] = false;
             $response['message'] = $ex->getMessage();
             return response($response, 500);
-    	}
+    	   }
    }
 
-   public function update(Request $request, Job $job){
-   		$job->update($request->all());
-
-   		return response()->json($job,200);
+   public function update(Request $request, Job $job)
+   {
+   		if($job->user_id == $request->user()->id || $request->user()->hasRole('admin')){
+            
+            $job->update($request->all());
+      
+      		return response()->json($job,200);
+         }else{
+            return response()->json([
+               'message' => 'User not allowed'
+            ],401);
+         }
    }
 
-   public function delete(Job $job){
-   		$job->delete();
-   		
-   		return $job;	
+   public function delete(Request $request, Job $job)
+   {
+   		if($job->user_id == $request->user()->id || $request->user()->hasRole('admin'))
+         {
+            $job->delete();
+   		   return $job;
+         }else{
+            return response()->json([
+               'message' => 'user not allowed'
+            ],401);
+         }
    }
 
 }
